apps/api/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
openai>=1.54,<2
pydantic==2.10.3
python-dotenv>=1.0,<2
pymongo==4.10.1
